requirements.txt
ipykernel
requests
sounddevice
numpy
openai
python-dotenv
websockets
elevenlabs
pyaudio
openai-agents[voice]
websocket-client
google-genai